VAM_AI/Six_Weeks/requirements.txt
requests
Pillow
gradio
together
